Move rule support to dedicated module
#433
