+++ build.gradle
@@ -50,7 +50,8 @@ ext {
 
 	jupiterProjects = [
 		'junit-jupiter-api',
-		'junit-jupiter-engine'
+		'junit-jupiter-engine',
+		'junit-jupiter-migration-support'
 	]
 
 	vintageProjects = [
@@ -67,6 +68,7 @@ ext {
 
 	cloverTestProjects = [
 		'junit-jupiter-engine',
+		'junit-jupiter-migration-support',
 		'junit-vintage-engine',
 		'junit-platform-surefire-provider',
 		'platform-tests'
