--- junit-jupiter-engine/src/main/java/org/junit/jupiter/engine/vintage/rulesupport/adapter/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Simple wrappers for JUnit 4 rules to overcome visibility limitations of the JUnit 4 implementations.
- */
-
-package org.junit.jupiter.engine.vintage.rulesupport.adapter;
