--- junit-jupiter-engine/src/main/java/org/junit/jupiter/engine/vintage/rulesupport/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Extensions which provide (limited) support for JUnit 4 rules within JUnit Jupiter.
- */
-
-package org.junit.jupiter.engine.vintage.rulesupport;
