+++ junit-jupiter-migration-support/src/main/java/org/junit/jupiter/migrationsupport/rules/adapter/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Simple wrappers for JUnit 4 rules to overcome visibility limitations of the JUnit 4 implementations.
+ */
+
+package org.junit.jupiter.migrationsupport.rules.adapter;
