+++ junit-jupiter-migration-support/src/main/java/org/junit/jupiter/migrationsupport/rules/member/RuleAnnotatedMethod.java
@@ -1,0 +1,25 @@
+/*
+ * Copyright 2015-2016 the original author or authors.
+ *
+ * All rights reserved. This program and the accompanying materials are
+ * made available under the terms of the Eclipse Public License v1.0 which
+ * accompanies this distribution and is available at
+ *
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
+
+package org.junit.jupiter.migrationsupport.rules.member;
+
+import java.lang.reflect.Method;
+
+import org.junit.platform.commons.util.ReflectionUtils;
+import org.junit.rules.TestRule;
+
+public class RuleAnnotatedMethod extends AbstractRuleAnnotatedMember {
+
+	public RuleAnnotatedMethod(Object testInstance, Method member) {
+		//no args
+		this.testRuleInstance = (TestRule) ReflectionUtils.invokeMethod(member, testInstance);
+	}
+
+}
