+++ junit-jupiter-migration-support/src/main/java/org/junit/jupiter/migrationsupport/rules/member/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Abstractions for members which can be targets of JUnit 4 rule annotations.
+ */
+
+package org.junit.jupiter.migrationsupport.rules.member;
