+++ junit-jupiter-migration-support/src/main/java/org/junit/jupiter/migrationsupport/rules/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Extensions which provide (limited) support for JUnit 4 rules within JUnit Jupiter.
+ */
+
+package org.junit.jupiter.migrationsupport.rules;
