+++ junit-jupiter-migration-support/src/main/resources/META-INF/services/org.junit.platform.engine.TestEngine
@@ -1,0 +1,1 @@
+org.junit.jupiter.engine.JupiterTestEngine
