+++ settings.gradle
@@ -3,6 +3,7 @@ rootProject.name = 'junit5'
 include 'documentation'
 include 'junit-jupiter-api'
 include 'junit-jupiter-engine'
+include 'junit-jupiter-migration-support'
 include 'junit-platform-commons'
 include 'junit-platform-console'
 include 'junit-platform-engine'
